Odczyt ciśnienia przy użyciu LPS25HB

+++ weather-station/Application/Inc/ui.h
@@ -12,7 +12,7 @@ enum UIStates {
 void NextUIState(void);
 void PrevUIState(void);
 void DrawCurrentUIState(uint8_t temperature, uint8_t humidity,
-		uint16_t pressure);
+		uint32_t pressure);
 void HandleBtnClick(uint8_t btn_pressed);
 
 #endif /* INC_UI_H_ */
